# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface EducationProblem {
  error: string;
}

export interface EducationPdfInput {
  file: Blob | File;
  /** @maxLength 100 */
  pageRange?: string;
}

export interface EducationSourcePage {
  /**
     * @minimum 1
     * @maximum 50
     * @nullable
     */
  pageNumber: number | null;
  /**
     * @minLength 1
     * @maxLength 250000
     */
  text: string;
}

export interface EducationExtraction {
  pages: EducationSourcePage[];
  totalPages: number;
  usedOcr: boolean;
  warnings: string[];
}

export type EducationQuizInputCount = typeof EducationQuizInputCount[keyof typeof EducationQuizInputCount];


export const EducationQuizInputCount = {
  NUMBER_5: 5,
  NUMBER_10: 10,
  NUMBER_20: 20,
  NUMBER_30: 30,
} as const;

export type EducationQuizInputQuestionTypesItem = typeof EducationQuizInputQuestionTypesItem[keyof typeof EducationQuizInputQuestionTypesItem];


export const EducationQuizInputQuestionTypesItem = {
  mcq: 'mcq',
  'true-false': 'true-false',
  'short-answer': 'short-answer',
  'fill-blank': 'fill-blank',
} as const;

export type EducationQuizInputDifficulty = typeof EducationQuizInputDifficulty[keyof typeof EducationQuizInputDifficulty];


export const EducationQuizInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
  mixed: 'mixed',
} as const;

export type EducationQuizInputLanguage = typeof EducationQuizInputLanguage[keyof typeof EducationQuizInputLanguage];


export const EducationQuizInputLanguage = {
  english: 'english',
  bangla: 'bangla',
  same: 'same',
} as const;

export interface EducationQuizInput {
  /**
     * @minItems 1
     * @maxItems 50
     */
  pages: EducationSourcePage[];
  count: EducationQuizInputCount;
  /** @minItems 1 */
  questionTypes: EducationQuizInputQuestionTypesItem[];
  difficulty: EducationQuizInputDifficulty;
  language: EducationQuizInputLanguage;
  /** @maxLength 200 */
  topicFocus?: string;
}

export type EducationFlashcardInputCount = typeof EducationFlashcardInputCount[keyof typeof EducationFlashcardInputCount];


export const EducationFlashcardInputCount = {
  NUMBER_10: 10,
  NUMBER_20: 20,
  NUMBER_40: 40,
  NUMBER_60: 60,
} as const;

export type EducationFlashcardInputStyle = typeof EducationFlashcardInputStyle[keyof typeof EducationFlashcardInputStyle];


export const EducationFlashcardInputStyle = {
  'term-definition': 'term-definition',
  'question-answer': 'question-answer',
  'concept-explanation': 'concept-explanation',
} as const;

export type EducationFlashcardInputLanguage = typeof EducationFlashcardInputLanguage[keyof typeof EducationFlashcardInputLanguage];


export const EducationFlashcardInputLanguage = {
  english: 'english',
  bangla: 'bangla',
  same: 'same',
} as const;

export interface EducationFlashcardInput {
  /**
     * @minItems 1
     * @maxItems 50
     */
  pages: EducationSourcePage[];
  count: EducationFlashcardInputCount;
  style: EducationFlashcardInputStyle;
  language: EducationFlashcardInputLanguage;
}

export type EducationQuestionType = typeof EducationQuestionType[keyof typeof EducationQuestionType];


export const EducationQuestionType = {
  mcq: 'mcq',
  'true-false': 'true-false',
  'short-answer': 'short-answer',
  'fill-blank': 'fill-blank',
} as const;

export type EducationQuestionDifficulty = typeof EducationQuestionDifficulty[keyof typeof EducationQuestionDifficulty];


export const EducationQuestionDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface EducationQuestion {
  /**
     * @minLength 1
     * @maxLength 100
     */
  id: string;
  type: EducationQuestionType;
  /**
     * @minLength 1
     * @maxLength 3000
     */
  question: string;
  /**
     * @maxItems 4
     * @items.maxLength 600
     */
  options: string[];
  /**
     * @minLength 1
     * @maxLength 3000
     */
  correctAnswer: string;
  /**
     * @minLength 1
     * @maxLength 3000
     */
  explanation: string;
  difficulty: EducationQuestionDifficulty;
  /**
     * @minimum 1
     * @maximum 50
     * @nullable
     */
  sourcePage: number | null;
}

export interface EducationQuiz {
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /**
     * @minItems 1
     * @maxItems 30
     */
  questions: EducationQuestion[];
}

export interface EducationCard {
  /**
     * @minLength 1
     * @maxLength 100
     */
  id: string;
  /**
     * @minLength 1
     * @maxLength 3000
     */
  front: string;
  /**
     * @minLength 1
     * @maxLength 3000
     */
  back: string;
  /**
     * @minimum 1
     * @maximum 50
     * @nullable
     */
  sourcePage: number | null;
}

export interface EducationDeck {
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /**
     * @minItems 1
     * @maxItems 100
     */
  cards: EducationCard[];
}

export type EducationExportInputKind = typeof EducationExportInputKind[keyof typeof EducationExportInputKind];


export const EducationExportInputKind = {
  questions: 'questions',
  answers: 'answers',
  cards: 'cards',
  quiz: 'quiz',
} as const;

export type EducationExportInputFormat = typeof EducationExportInputFormat[keyof typeof EducationExportInputFormat];


export const EducationExportInputFormat = {
  pdf: 'pdf',
  docx: 'docx',
} as const;

export interface EducationExportInput {
  kind: EducationExportInputKind;
  format: EducationExportInputFormat;
  quiz?: EducationQuiz;
  deck?: EducationDeck;
}

export type EducationEventEvent = typeof EducationEventEvent[keyof typeof EducationEventEvent];


export const EducationEventEvent = {
  upload: 'upload',
  generate: 'generate',
  export: 'export',
  quiz_completed: 'quiz_completed',
} as const;

export type EducationEventTool = typeof EducationEventTool[keyof typeof EducationEventTool];


export const EducationEventTool = {
  'quiz-generator': 'quiz-generator',
  flashcards: 'flashcards',
} as const;

export interface EducationEvent {
  event: EducationEventEvent;
  tool: EducationEventTool;
}

export interface HealthStatus {
  status: string;
}

/**
 * Friendly processing error
 */
export type EducationErrorResponse = EducationProblem;