# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
paths:
  /education/extract:
    post:
      operationId: extractEducationPdf
      tags: [education]
      summary: Read a temporary PDF, using OCR for pages without text
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/EducationPdfInput"
      responses:
        "200":
          description: Extracted pages; nothing is stored on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EducationExtraction"
        "400":
          $ref: "#/components/responses/EducationError"
        "413":
          $ref: "#/components/responses/EducationError"
        "429":
          $ref: "#/components/responses/EducationError"
  /education/quiz:
    post:
      operationId: generateEducationQuiz
      tags: [education]
      summary: Generate a document-grounded quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationQuizInput"
      responses:
        "200":
          description: Validated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EducationQuiz"
        "400":
          $ref: "#/components/responses/EducationError"
        "429":
          $ref: "#/components/responses/EducationError"
        "503":
          $ref: "#/components/responses/EducationError"
  /education/flashcards:
    post:
      operationId: generateEducationFlashcards
      tags: [education]
      summary: Generate a document-grounded flashcard deck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationFlashcardInput"
      responses:
        "200":
          description: Validated flashcard deck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EducationDeck"
        "400":
          $ref: "#/components/responses/EducationError"
        "429":
          $ref: "#/components/responses/EducationError"
        "503":
          $ref: "#/components/responses/EducationError"
  /education/export:
    post:
      operationId: exportEducation
      tags: [education]
      summary: Export an edited quiz or printable deck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationExportInput"
      responses:
        "200":
          description: PDF or DOCX file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/EducationError"
        "503":
          $ref: "#/components/responses/EducationError"
  /education/events:
    post:
      operationId: recordEducationEvent
      tags: [education]
      summary: First-party activity counts, without document content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationEvent"
      responses:
        "204":
          description: Recorded
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
components:
  responses:
    EducationError:
      description: Friendly processing error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EducationProblem"
  schemas:
    EducationProblem:
      type: object
      required: [error]
      properties:
        error: { type: string }
    EducationPdfInput:
      type: object
      required: [file]
      properties:
        file: { type: string, format: binary }
        pageRange: { type: string, maxLength: 100 }
    EducationSourcePage:
      type: object
      required: [pageNumber, text]
      properties:
        pageNumber: { type: [integer, "null"], minimum: 1, maximum: 50 }
        text: { type: string, minLength: 1, maxLength: 250000 }
    EducationExtraction:
      type: object
      required: [pages, totalPages, usedOcr, warnings]
      properties:
        pages:
          type: array
          items: { $ref: "#/components/schemas/EducationSourcePage" }
        totalPages: { type: integer }
        usedOcr: { type: boolean }
        warnings: { type: array, items: { type: string } }
    EducationQuizInput:
      type: object
      required: [pages, count, questionTypes, difficulty, language]
      properties:
        pages:
          type: array
          minItems: 1
          maxItems: 50
          items: { $ref: "#/components/schemas/EducationSourcePage" }
        count: { type: integer, enum: [5, 10, 20, 30] }
        questionTypes:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, enum: [mcq, true-false, short-answer, fill-blank] }
        difficulty: { type: string, enum: [easy, medium, hard, mixed] }
        language: { type: string, enum: [english, bangla, same] }
        topicFocus: { type: string, maxLength: 200 }
    EducationFlashcardInput:
      type: object
      required: [pages, count, style, language]
      properties:
        pages:
          type: array
          minItems: 1
          maxItems: 50
          items: { $ref: "#/components/schemas/EducationSourcePage" }
        count: { type: integer, enum: [10, 20, 40, 60] }
        style: { type: string, enum: [term-definition, question-answer, concept-explanation] }
        language: { type: string, enum: [english, bangla, same] }
    EducationQuestion:
      type: object
      required: [id, type, question, options, correctAnswer, explanation, difficulty, sourcePage]
      properties:
        id: { type: string, minLength: 1, maxLength: 100 }
        type: { type: string, enum: [mcq, true-false, short-answer, fill-blank] }
        question: { type: string, minLength: 1, maxLength: 3000 }
        options: { type: array, maxItems: 4, items: { type: string, maxLength: 600 } }
        correctAnswer: { type: string, minLength: 1, maxLength: 3000 }
        explanation: { type: string, minLength: 1, maxLength: 3000 }
        difficulty: { type: string, enum: [easy, medium, hard] }
        sourcePage: { type: [integer, "null"], minimum: 1, maximum: 50 }
    EducationQuiz:
      type: object
      required: [title, questions]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        questions:
          type: array
          minItems: 1
          maxItems: 30
          items: { $ref: "#/components/schemas/EducationQuestion" }
    EducationCard:
      type: object
      required: [id, front, back, sourcePage]
      properties:
        id: { type: string, minLength: 1, maxLength: 100 }
        front: { type: string, minLength: 1, maxLength: 3000 }
        back: { type: string, minLength: 1, maxLength: 3000 }
        sourcePage: { type: [integer, "null"], minimum: 1, maximum: 50 }
    EducationDeck:
      type: object
      required: [title, cards]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        cards:
          type: array
          minItems: 1
          maxItems: 100
          items: { $ref: "#/components/schemas/EducationCard" }
    EducationExportInput:
      type: object
      required: [kind, format]
      properties:
        kind: { type: string, enum: [questions, answers, cards, quiz] }
        format: { type: string, enum: [pdf, docx] }
        quiz: { $ref: "#/components/schemas/EducationQuiz" }
        deck: { $ref: "#/components/schemas/EducationDeck" }
    EducationEvent:
      type: object
      required: [event, tool]
      properties:
        event: { type: string, enum: [upload, generate, export, quiz_completed] }
        tool: { type: string, enum: [quiz-generator, flashcards] }
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status