# lib/api-zod/src/generated/types/educationExportInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EducationDeck } from './educationDeck';
import type { EducationExportInputFormat } from './educationExportInputFormat';
import type { EducationExportInputKind } from './educationExportInputKind';
import type { EducationQuiz } from './educationQuiz';

export interface EducationExportInput {
  kind: EducationExportInputKind;
  format: EducationExportInputFormat;
  quiz?: EducationQuiz;
  deck?: EducationDeck;
}