# lib/api-zod/src/generated/types/educationFlashcardInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EducationFlashcardInputCount } from './educationFlashcardInputCount';
import type { EducationFlashcardInputLanguage } from './educationFlashcardInputLanguage';
import type { EducationFlashcardInputStyle } from './educationFlashcardInputStyle';
import type { EducationSourcePage } from './educationSourcePage';

export interface EducationFlashcardInput {
  /**
     * @minItems 1
     * @maxItems 50
     */
  pages: EducationSourcePage[];
  count: EducationFlashcardInputCount;
  style: EducationFlashcardInputStyle;
  language: EducationFlashcardInputLanguage;
}