# lib/api-zod/src/generated/types/educationPdfInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EducationPdfInput {
  file: Blob | File;
  /** @maxLength 100 */
  pageRange?: string;
}