# lib/api-zod/src/generated/types/educationQuestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EducationQuestionDifficulty } from './educationQuestionDifficulty';
import type { EducationQuestionType } from './educationQuestionType';

export interface EducationQuestion {
  /**
     * @minLength 1
     * @maxLength 100
     */
  id: string;
  type: EducationQuestionType;
  /**
     * @minLength 1
     * @maxLength 3000
     */
  question: string;
  /**
     * @maxItems 4
     * @items.maxLength 600
     */
  options: string[];
  /**
     * @minLength 1
     * @maxLength 3000
     */
  correctAnswer: string;
  /**
     * @minLength 1
     * @maxLength 3000
     */
  explanation: string;
  difficulty: EducationQuestionDifficulty;
  /**
     * @minimum 1
     * @maximum 50
     * @nullable
     */
  sourcePage: number | null;
}