# lib/api-zod/src/generated/types/educationSourcePage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EducationSourcePage {
  /**
     * @minimum 1
     * @maximum 50
     * @nullable
     */
  pageNumber: number | null;
  /**
     * @minLength 1
     * @maxLength 250000
     */
  text: string;
}